fix: Виправлено створення symlink

+++ examples/scripts/link_lib.py
@@ -10,9 +10,11 @@ lib_dst = os.path.join(libdeps_dir, lib_name)
 os.makedirs(libdeps_dir, exist_ok=True)
 
 if os.path.islink(lib_dst):
-    os.unlink(lib_dst)
+    if os.path.realpath(lib_dst) != os.path.realpath(lib_src):
+        os.unlink(lib_dst)
+        os.symlink(lib_src, lib_dst)
 elif os.path.exists(lib_dst):
     shutil.rmtree(lib_dst)
-
-os.symlink(lib_src, lib_dst)
-print(f"[symlink] {lib_src} -> {lib_dst}")
+    os.symlink(lib_src, lib_dst)
+else:
+    os.symlink(lib_src, lib_dst)
